chore: moved logic outside index.js file
