+++ src/card/lib/card-utils.js
@@ -3,7 +3,8 @@
 
 import { values } from '@krakenjs/belter';
 
-import type { InputState, FieldValidity, ExtraFields, CardType, ParsedCardType, PaymentSourceCardInput, ReformattedPaymentSourceCardInput } from '../types';
+import type { PaymentSourceInput } from "../../api/vault";
+import type { InputState, FieldValidity, Card, ExtraFields, CardType, ParsedCardType, PaymentSourceCardInput, ReformattedPaymentSourceCardInput } from '../types';
 import {
     CARD_ERRORS,
     CARD_FIELD_TYPE,
@@ -318,6 +319,61 @@ export function parsedCardType(potentialCardTypes: $ReadOnlyArray<CardType>) : $
 export function getContext(win : Object) : string {
     return win.xprops?.parent?.uid || win.xprops?.uid;
 }
+
+export function cardExpiryToPaymentSourceExpiry(dateString: string): string {
+    if (!dateString || typeof dateString !== "string") {
+        throw new Error(`can not convert invalid expiry date: ${dateString}`);
+    }
+
+    // "2020-12"
+    const YYYYmmRegex = "^[0-9]{4}-([1-9]|0[1-9]|1[0-2])$";
+    // 12/20 OR 12/2020
+    const mmYYYYRegex = "^([1-9]|0[1-9]|1[0-2])/?([0-9]{4}|[0-9]{2})$";
+
+    if (dateString.match(YYYYmmRegex)) {
+        return dateString;
+    }
+
+    if (dateString.match(mmYYYYRegex)) {
+        const [monthString, yearString] = dateString.split("/");
+
+        const formattedYearString =
+            yearString.length === 2 ? `20${yearString}` : yearString;
+        const formattedMonthString =
+            monthString.length === 1 ? `0${monthString}` : monthString;
+
+        return `${formattedYearString}-${formattedMonthString}`;
+    }
+
+    throw new Error(`can not convert invalid expiry date: ${dateString}`);
+};
+
+export function convertCardToPaymentSource(card: Card, extraFields?: ExtraFields): PaymentSourceInput {
+    const paymentSource = {
+        card: {
+            number: card.number,
+            securityCode: card.cvv,
+            expiry: cardExpiryToPaymentSourceExpiry(card.expiry),
+        },
+    };
+
+    if (extraFields && Object.keys(extraFields).length !== 0) {
+        // $FlowIssue
+        paymentSource.card.billingAddress = extraFields.billingAddress;
+    }
+
+    if (card.name) {
+        // $FlowIssue
+        paymentSource.card.name = card.name;
+    }
+
+    if (card.postalCode) {
+        // $FlowIssue
+        paymentSource.card.billingAddress = { postalCode: card.postalCode };
+    }
+
+    return paymentSource;
+};
 
 export function kebabToCamelCase(field: string): string {
     const camelCase = field.split("-");
